feat(student) one busuines logic added

+++ MyAspNetProject/Models/Domain/Klass.cs
@@ -2,6 +2,6 @@ namespace MyAspNetProject.Models.Domain;
 
 public class Klass: BaseModel
 {
-    public short Year { get; set; }
+    public int Year { get; set; }
     public required string Group { get; set; }
 }
